+++ src/elite/test/EliteSharpLib.Fakes/FakeEliteDraw.cs
@@ -36,7 +36,7 @@ internal class FakeEliteDraw : IEliteDraw
 
     public float Top => 0;
 
-    public List<(Vector2[] Points, uint FaceColor, float AverageZ)> DrawnPolygons { get; } = [];
+    public List<(Vector2[] Points, uint FaceColor, float Z)> DrawnPolygons { get; } = [];
 
     public void DrawBorder()
     {
@@ -50,8 +50,8 @@ internal class FakeEliteDraw : IEliteDraw
     {
     }
 
-    public void DrawPolygonFilled(Vector2[] points, uint faceColor, float averageZ)
-        => DrawnPolygons.Add((points, faceColor, averageZ));
+    public void DrawPolygonFilled(Vector2[] points, uint faceColor, float z)
+        => DrawnPolygons.Add((points, faceColor, z));
 
     public void DrawTextPretty(Vector2 position, float width, string text)
     {